ENG-123 Eliminate pointless interface
