+++ src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -43,7 +43,7 @@
   <beans:bean class="org.ambraproject.wombat.config.RootConfiguration" />
 
   <beans:bean id="runtimeConfiguration"
-        class="org.ambraproject.wombat.config.RuntimeConfigurationImpl">
+        class="org.ambraproject.wombat.config.RuntimeConfiguration">
     <beans:property name="solrUrl" value="${SOLR_URL:}" />
     <beans:property name="nedUrl" value="${NED_URL:}" />
     <beans:property name="casUrl" value="${CAS_URL:}" />
